midterm

+++ src/design/Employee.java
@@ -18,7 +18,7 @@ public interface Employee {
     public void assignDepartment();
 	
 	//calculate employee salary
-	public int calculateSalary();
+	public double calculateSalary();
 	
 	//employee benefit
 	public void benefitLayout();
@@ -26,6 +26,9 @@ public interface Employee {
 	public void employeeRank();
 
 	public void employeeGender();
+
+
+
